Extract SelectedFilesCard to its own component
